+++ example/lib/shared/util/player_sprite_sheet.dart
@@ -39,9 +39,7 @@ class PlayerSpriteSheet {
 
   static SimpleDirectionAnimation get simpleDirectionAnimation =>
       SimpleDirectionAnimation(
-        idleLeft: idleLeft,
         idleRight: idleRight,
-        runLeft: runLeft,
         runRight: runRight,
       );
 }